+++ frontend/src/metabase/auth/actions.ts
@@ -2,7 +2,7 @@ import { type UnknownAction, createAction } from "@reduxjs/toolkit";
 import { getIn } from "icepick";
 import { push } from "react-router-redux";
 
-import { loadLocalization } from "metabase/api/load-localization";
+import { loadLocalization } from "metabase/api/localization";
 import { openNavbar } from "metabase/redux/app";
 import { refreshSiteSettings } from "metabase/redux/settings";
 import { clearCurrentUser, refreshCurrentUser } from "metabase/redux/user";
