+++ frontend/src/metabase/embedding/LocaleProvider.tsx
@@ -7,7 +7,7 @@ import {
 } from "react";
 
 import { setLocaleHeader } from "metabase/api/legacy-client";
-import { loadLocalization } from "metabase/api/load-localization";
+import { loadLocalization } from "metabase/api/localization";
 import { useSetting } from "metabase/common/hooks";
 import { isEmbeddingSdk } from "metabase/embedding-sdk/config";
 import { DatesProvider } from "metabase/ui/components/theme/DatesProvider/DatesProvider";
